# Conflicts:
#	Engine/Modules/Core/BlackSmith.php

+++ Engine/Modules/TemplateEngine/Services/JsAssetService.php
@@ -31,16 +31,22 @@ class JsAssetService extends BaseAssetService
      * @throws \Doctrine\ORM\OptimisticLockException
      * @throws \Oforge\Engine\Modules\Core\Exceptions\ServiceNotFoundException
      */
-    public function build(string $scope = parent::DEFAULT_SCOPE): string
+    public function build(string $scope = TemplateAssetService::DEFAULT_SCOPE): string
     {
         parent::build();
         $dirs = $this->getAssetsDirectories();
         
         $fileName = "scripts." . bin2hex(openssl_random_pseudo_bytes(16));
-        
-        $output = Statics::ASSET_CACHE_DIR . DIRECTORY_SEPARATOR . $fileName;
+
+        $folder = Statics::ASSET_CACHE_DIR . DIRECTORY_SEPARATOR . $scope . DIRECTORY_SEPARATOR . $this->key;
+        $fullFolder = ROOT_PATH . $folder;
+        $output = $folder . DIRECTORY_SEPARATOR . $fileName;
         $outputFull = ROOT_PATH . $output;
-        
+
+        if(!file_exists($fullFolder) || (file_exists($fullFolder) && !is_dir($fullFolder))) {
+            mkdir($fullFolder, 0750, true);
+        }
+
         //iterate over all plugins, current theme and base theme
         foreach ($dirs as $dir) {
             $folder = $dir . DIRECTORY_SEPARATOR . $scope . DIRECTORY_SEPARATOR . Statics::ASSETS_DIR  . DIRECTORY_SEPARATOR . Statics::ASSETS_JS . DIRECTORY_SEPARATOR;
@@ -50,7 +56,7 @@ class JsAssetService extends BaseAssetService
                     while(!feof($file)) {
                         $line = trim(fgets($file));
 
-                        if(file_exists($folder . $line)) {
+                        if(strlen($line) > 0 && file_exists($folder . $line) ) {
                             file_put_contents($outputFull . ".js", file_get_contents($folder . $line), FILE_APPEND);
                         }
                     }
@@ -59,11 +65,13 @@ class JsAssetService extends BaseAssetService
             }
         }
 
+
         $minifier = new JS($outputFull . ".js");
         $minifier->minify($outputFull . ".min.js");
 
         $this->store->set($this->getAccessKey($scope), $output . ".min.js");
-        $this->removeOldAssets($fileName ,".js");
+        $this->removeOldAssets($fullFolder, $fileName ,".js");
+
         return $output . ".min.js";
     }
 }
